requirements.txt
fastapi
uvicorn
load_dotenv
SQLAlchemy
psycopg2-binary
pydantic_settings